+++ Contrib/at67/compiler.cpp
@@ -150,7 +150,7 @@ namespace Compiler
         }
 
         // Write output .vasm file
-        for(size_t i=0; i<_output.size(); i++)
+        for(int i=0; i<int(_output.size()); i++)
         {
             outfile.write((char *)_output[i].c_str(), _output[i].size());
             if(outfile.bad() || outfile.fail())
@@ -296,7 +296,7 @@ namespace Compiler
         if(codeLine._code.size() < 2  ||  equals >= codeLine._code.size()) return VarNotFound;
 
         // Check all tokens individually, don't just do a find as a var may exist with a reserved keyword embedded within it
-        for(size_t i=0; i<codeLine._tokens.size(); i++)
+        for(int i=0; i<int(codeLine._tokens.size()); i++)
         {
             std::string token = codeLine._tokens[i];
             Expression::stripWhitespace(token);
@@ -350,7 +350,7 @@ namespace Compiler
         if(codeLine._code.size() < 2  ||  equals >= codeLine._code.size() - 2  ||  codeLine._tokens.size() < 2) return VarError;
 
         // Check all tokens individually, don't just do a find as a str name may exist with a reserved keyword embedded within it
-        for(size_t i=0; i<codeLine._tokens.size(); i++)
+        for(int i=0; i<int(codeLine._tokens.size()); i++)
         {
             std::string token = codeLine._tokens[i];
             Expression::stripWhitespace(token);
@@ -426,7 +426,7 @@ namespace Compiler
         std::vector<std::string> tokens = Expression::tokenise(stripped, "-+/*%&<>=();, ", false);
 
         // Check for keywords
-        for(size_t i=0; i<tokens.size(); i++)
+        for(int i=0; i<int(tokens.size()); i++)
         {
             std::string token = tokens[i];
             Expression::strToUpper(token);
@@ -438,7 +438,7 @@ namespace Compiler
         }
 
         // Check for functions
-        for(size_t i=0; i<tokens.size(); i++)
+        for(int i=0; i<int(tokens.size()); i++)
         {
             std::string token = tokens[i];
             Expression::strToUpper(token);
@@ -450,7 +450,7 @@ namespace Compiler
         }
 
         // Check for string keywords
-        for(size_t i=0; i<tokens.size(); i++)
+        for(int i=0; i<int(tokens.size()); i++)
         {
             std::string token = tokens[i];
             Expression::strToUpper(token);
@@ -465,7 +465,7 @@ namespace Compiler
         }
 
         // Check for consts
-        for(size_t i=0; i<tokens.size(); i++)
+        for(int i=0; i<int(tokens.size()); i++)
         {
             constIndex = findConst(tokens[i]);
             if(constIndex != -1  &&  tokens[i][0] != '@') // 'address of' operator returns numbers
@@ -484,7 +484,7 @@ namespace Compiler
         }
 
         // Check for vars
-        for(size_t i=0; i<tokens.size(); i++)
+        for(int i=0; i<int(tokens.size()); i++)
         {
             varIndex = findVar(tokens[i]);
             if(varIndex != -1  &&  tokens[i][0] != '@') // 'address of' operator returns numbers
@@ -502,7 +502,7 @@ namespace Compiler
         }
 
         // Check for string vars
-        for(size_t i=0; i<tokens.size(); i++)
+        for(int i=0; i<int(tokens.size()); i++)
         {
             strIndex = findStr(tokens[i]);
             if(strIndex != -1  &&  _stringVars[strIndex]._constant == false  &&  tokens[i][0] != '@') // 'address of' operator returns numbers)
@@ -519,7 +519,7 @@ namespace Compiler
             if(stripped.find_first_of("-+/*%&<>=") != std::string::npos) expressionType |= Expression::HasOperators;
             std::string mod = stripped;
             Expression::strToUpper(mod);
-            for(size_t i=0; i<Keywords::getOperators().size(); i++)
+            for(int i=0; i<int(Keywords::getOperators().size()); i++)
             {
                 if(mod.find(Keywords::getOperators()[i]) != std::string::npos) expressionType |= Expression::HasOperators;
             }
@@ -603,7 +603,7 @@ namespace Compiler
             std::string macroLine = (commentStart != std::string::npos) ? _macroLines[i].substr(0, commentStart) : _macroLines[i];
             std::vector<std::string> tokens = Expression::tokeniseLine(macroLine);
 
-            for(size_t j=0; j<tokens.size(); j++)
+            for(int j=0; j<int(tokens.size()); j++)
             {
                 Expression::stripWhitespace(tokens[j]);
                 Expression::strToUpper(tokens[j]);
@@ -611,7 +611,7 @@ namespace Compiler
             }
 
             // Check for nested macros
-            for(size_t j=0; j<tokens.size(); j++)
+            for(int j=0; j<int(tokens.size()); j++)
             {
                 if(findMacroText(tokens[j], text)) return true;
             }
@@ -635,7 +635,7 @@ namespace Compiler
             std::vector<std::string> tokens = Expression::tokeniseLine(macroLine);
 
             int opcodeSize = 0;
-            for(size_t j=0; j<tokens.size(); j++)
+            for(int j=0; j<int(tokens.size()); j++)
             {
                 opcodeSize = Assembler::getAsmOpcodeSize(tokens[j]);
                 if(opcodeSize)
@@ -648,7 +648,7 @@ namespace Compiler
             // Check for nested macros
             if(opcodeSize == 0)
             {
-                for(size_t j=0; j<tokens.size(); j++)
+                for(int j=0; j<int(tokens.size()); j++)
                 {
                     opcodeSize = getMacroSize(tokens[j]);
                     if(opcodeSize)
@@ -1081,7 +1081,7 @@ namespace Compiler
 
     bool isGosubLabel(const std::string& label)
     {
-        for(size_t i=0; i<_gosubLabels.size(); i++)
+        for(int i=0; i<int(_gosubLabels.size()); i++)
         {
             if(_gosubLabels[i] == label) return true;
         }
@@ -1175,7 +1175,7 @@ namespace Compiler
             bool validLabel = true;
             if(labelName.size())
             {
-                for(size_t i=0; i<labelName.size(); i++)
+                for(int i=0; i<int(labelName.size()); i++)
                 {
                     if(!(labelName[i] == '_')  &&  !isalnum((unsigned char)labelName[i]))
                     {
@@ -1234,7 +1234,7 @@ namespace Compiler
             {
                 std::vector<size_t> offsets;
                 std::vector<std::string> tokens = Expression::tokenise(*it, ' ', offsets, false, false);
-                for(size_t i=0; i<tokens.size(); i++)
+                for(int i=0; i<int(tokens.size()); i++)
                 {
                     Expression::stripWhitespace(tokens[i]);
                 }
@@ -1972,7 +1972,7 @@ namespace Compiler
 
                 // Source string addresses, (extra 0x0000 delimiter used by VASM runtime)
                 std::vector<uint16_t> strAddrs(tokens.size() + 1, 0x0000);
-                for(size_t i=0; i<tokens.size(); i++)
+                for(int i=0; i<int(tokens.size()); i++)
                 {
                     Expression::stripNonStringWhitespace(tokens[i]);
                     strAddrs[i] = getStringSrcAddr(tokens[i]);
@@ -2161,7 +2161,7 @@ namespace Compiler
         // Tokenise and parse multi-statement lines
         StatementResult statementResult = StatementSuccess;
         std::vector<std::string> tokens = Expression::tokenise(code, ':', false);
-        for(size_t j=0; j<tokens.size(); j++)
+        for(int j=0; j<int(tokens.size()); j++)
         {
             createCodeLine(tokens[j], 0, _codeLines[codeLineIndex]._labelIndex, -1, Expression::Int16Both, false, codeline);
             createCodeVar(codeline, codeLineIndex, varIndex);
@@ -2178,7 +2178,7 @@ namespace Compiler
 
     void addLabelToJumpCC(std::vector<VasmLine>& vasm, const std::string& label)
     {
-        for(size_t i=0; i<vasm.size(); i++)
+        for(int i=0; i<int(vasm.size()); i++)
         {
             if(vasm[i]._code.substr(0, sizeof("Jump")-1) == "Jump"  ||  vasm[i]._code.substr(0, sizeof("B")-1) == "B")
             {
@@ -2190,7 +2190,7 @@ namespace Compiler
 
     void addLabelToJump(std::vector<VasmLine>& vasm, const std::string& label)
     {
-        for(size_t i=0; i<vasm.size(); i++)
+        for(int i=0; i<int(vasm.size()); i++)
         {
             if(Assembler::getUseOpcodeCALLI())
             {
@@ -2231,7 +2231,7 @@ namespace Compiler
     {
         uint16_t match;
         Expression::stringToU16(name.substr(name.size() - 6, 6), match);
-        for(size_t i=0; i<_discardedLabels.size(); i++)
+        for(int i=0; i<int(_discardedLabels.size()); i++)
         {
             if(_discardedLabels[i]._address == match) _discardedLabels[i]._address = address;
         }
@@ -2257,7 +2257,7 @@ namespace Compiler
         if(_codeLines.size())
         {
             bool foundEnd = false;
-            for(size_t i=0; i<_codeLines[_codeLines.size() - 1]._tokens.size(); i++)
+            for(int i=0; i<int(_codeLines[_codeLines.size() - 1]._tokens.size()); i++)
             {
                 std::string token = _codeLines[_codeLines.size() - 1]._tokens[i];
                 Expression::strToUpper(token);
@@ -2311,7 +2311,7 @@ namespace Compiler
         _output.push_back("; Labels\n");
 
         // BASIC labels
-        for(size_t i=0; i<_labels.size(); i++)
+        for(int i=0; i<int(_labels.size()); i++)
         {
             std::string address = Expression::wordToHexString(_labels[i]._address);
             _output.push_back(_labels[i]._output + "EQU" + std::string(OPCODE_TRUNC_SIZE - 3, ' ') + address + "\n");
@@ -2319,9 +2319,9 @@ namespace Compiler
         }
 
         // Internal labels
-        for(size_t i=0; i<_codeLines.size(); i++)
+        for(int i=0; i<int(_codeLines.size()); i++)
         {
-            for(size_t j=0; j<_codeLines[i]._vasm.size(); j++)
+            for(int j=0; j<int(_codeLines[i]._vasm.size()); j++)
             {
                 std::string internalLabel = _codeLines[i]._vasm[j]._internalLabel;
                 if(internalLabel.size())
@@ -2335,12 +2335,12 @@ namespace Compiler
         }
 
         // Check for label conflicts
-        for(size_t i=0; i<_codeLines.size(); i++)
+        for(int i=0; i<int(_codeLines.size()); i++)
         {
-            for(size_t j=0; j<_codeLines[i]._vasm.size(); j++)
+            for(int j=0; j<int(_codeLines[i]._vasm.size()); j++)
             {
                 // BASIC label conflict
-                for(size_t k=0; k<_internalLabels.size(); k++)
+                for(int k=0; k<int(_internalLabels.size()); k++)
                 {
                     int labelIndex = findLabel(_internalLabels[k]._address);
                     if(labelIndex >= 0)
@@ -2357,7 +2357,7 @@ namespace Compiler
                     }
 
                     // Discarded internal label
-                    for(size_t l=0; l<_discardedLabels.size(); l++)
+                    for(int l=0; l<int(_discardedLabels.size()); l++)
                     {
                         // Match on unique address embedded within names or the real address
                         std::string internalName = _internalLabels[k]._name.substr(_internalLabels[k]._name.size() - 4, 4);
@@ -2378,7 +2378,7 @@ namespace Compiler
     {
         _output.push_back("; Constants\n");
 
-        for(size_t i=0; i<_constants.size(); i++)
+        for(int i=0; i<int(_constants.size()); i++)
         {
             int16_t data = _constants[i]._data;
             std::string name = _constants[i]._name;
@@ -2409,7 +2409,7 @@ namespace Compiler
     {
         _output.push_back("; Variables\n");
 
-        for(size_t i=0; i<_integerVars.size(); i++)
+        for(int i=0; i<int(_integerVars.size()); i++)
         {
             switch(_integerVars[i]._varType)
             {
@@ -2445,7 +2445,7 @@ namespace Compiler
     {
         _output.push_back("; Strings\n");
 
-        for(size_t i=0; i<_stringVars.size(); i++)
+        for(int i=0; i<int(_stringVars.size()); i++)
         {
             _output.push_back(_stringVars[i]._output + "EQU" + std::string(OPCODE_TRUNC_SIZE - 3, ' ') + Expression::wordToHexString(_stringVars[i]._address) + "\n");
             _output.push_back(_stringVars[i]._output + "DB" + std::string(OPCODE_TRUNC_SIZE - 2, ' ') + std::to_string(_stringVars[i]._size) + " '" + _stringVars[i]._text + "' 0\n");
@@ -2458,13 +2458,13 @@ namespace Compiler
     {
         // Create def byte data
         _output.push_back("; Define Bytes\n");
-        for(size_t i=0; i<_defDataBytes.size(); i++)
+        for(int i=0; i<int(_defDataBytes.size()); i++)
         {
             std::string defName = "def_bytes_" + Expression::wordToHexString(_defDataBytes[i]._address);
             _output.push_back(defName + std::string(LABEL_TRUNC_SIZE - defName.size(), ' ') + "EQU" + std::string(OPCODE_TRUNC_SIZE - 3, ' ') + Expression::wordToHexString(_defDataBytes[i]._address) + "\n");
             
             std::string dbString = defName + std::string(LABEL_TRUNC_SIZE - defName.size(), ' ') + "DB" + std::string(OPCODE_TRUNC_SIZE - 2, ' ');
-            for(size_t j=0; j<_defDataBytes[i]._data.size(); j++)
+            for(int j=0; j<int(_defDataBytes[i]._data.size()); j++)
             {
                 dbString += std::to_string(_defDataBytes[i]._data[j]) + " ";
             }
@@ -2474,13 +2474,13 @@ namespace Compiler
 
         // Create def word data
         _output.push_back("; Define Words\n");
-        for(size_t i=0; i<_defDataWords.size(); i++)
+        for(int i=0; i<int(_defDataWords.size()); i++)
         {
             std::string defName = "def_words_" + Expression::wordToHexString(_defDataWords[i]._address);
             _output.push_back(defName + std::string(LABEL_TRUNC_SIZE - defName.size(), ' ') + "EQU" + std::string(OPCODE_TRUNC_SIZE - 3, ' ') + Expression::wordToHexString(_defDataWords[i]._address) + "\n");
             
             std::string dwString = defName + std::string(LABEL_TRUNC_SIZE - defName.size(), ' ') + "DW" + std::string(OPCODE_TRUNC_SIZE - 2, ' ');
-            for(size_t j=0; j<_defDataWords[i]._data.size(); j++)
+            for(int j=0; j<int(_defDataWords[i]._data.size()); j++)
             {
                 dwString += std::to_string(_defDataWords[i]._data[j]) + " ";
             }
@@ -2500,7 +2500,7 @@ namespace Compiler
         {
             std::vector<uint16_t> numericLabels;
             std::vector<uint16_t> numericAddresses;
-            for(size_t i=0; i<_labels.size(); i++)
+            for(int i=0; i<int(_labels.size()); i++)
             {
                 if(_labels[i]._numeric)
                 {
@@ -2531,7 +2531,7 @@ namespace Compiler
                 _output.push_back(lutName + std::string(LABEL_TRUNC_SIZE - lutName.size(), ' ') + "EQU" + std::string(OPCODE_TRUNC_SIZE - 3, ' ') + Expression::wordToHexString(lutAddress) + "\n");
             
                 std::string dbString = lutName + std::string(LABEL_TRUNC_SIZE - lutName.size(), ' ') + "DW" + std::string(OPCODE_TRUNC_SIZE - 2, ' ');
-                for(size_t i=0; i<numericLabels.size(); i++)
+                for(int i=0; i<int(numericLabels.size()); i++)
                 {
                     dbString += std::to_string(numericLabels[i]) + " ";
                 }
@@ -2548,7 +2548,7 @@ namespace Compiler
                 _output.push_back(lutName + std::string(LABEL_TRUNC_SIZE - lutName.size(), ' ') + "EQU" + std::string(OPCODE_TRUNC_SIZE - 3, ' ') + Expression::wordToHexString(lutAddress) + "\n");
             
                 std::string dwString = lutName + std::string(LABEL_TRUNC_SIZE - lutName.size(), ' ') + "DW" + std::string(OPCODE_TRUNC_SIZE - 2, ' ');
-                for(size_t i=0; i<numericAddresses.size(); i++)
+                for(int i=0; i<int(numericAddresses.size()); i++)
                 {
                     dwString += Expression::wordToHexString(numericAddresses[i]) + " ";
                 }
@@ -2557,7 +2557,7 @@ namespace Compiler
         }
 
         // ON GOTO GOSUB label LUTs
-        for(size_t i=0; i<_codeLines.size(); i++)
+        for(int i=0; i<int(_codeLines.size()); i++)
         {
             // Output LUT if it exists
             int lutSize = int(_codeLines[i]._onGotoGosubLut._lut.size());
@@ -2581,7 +2581,7 @@ namespace Compiler
         }
 
         // String concatenation LUTs
-        for(size_t i=0; i<_codeLines.size(); i++)
+        for(int i=0; i<int(_codeLines.size()); i++)
         {
             // Output LUT if it exists
             int lutSize = int(_codeLines[i]._strConcatLut._lut.size());
@@ -2602,7 +2602,7 @@ namespace Compiler
         }
 
         // INPUT LUTs
-        for(size_t i=0; i<_codeLines.size(); i++)
+        for(int i=0; i<int(_codeLines.size()); i++)
         {
             // Output varsLUT if it exists
             int varsSize = int(_codeLines[i]._inputLut._varsLut.size());
@@ -2734,7 +2734,7 @@ namespace Compiler
 
         _output.push_back("; Code\n");
 
-        for(size_t i=0; i<_codeLines.size(); i++)
+        for(int i=0; i<int(_codeLines.size()); i++)
         {
             int labelIndex = _codeLines[i]._labelIndex;
 
@@ -2745,7 +2745,7 @@ namespace Compiler
                 std::string basicLabel = (labelIndex >= 0) ? _labels[labelIndex]._output : "";
 
                 // Vasm code
-                for(size_t j=0; j<_codeLines[i]._vasm.size(); j++)
+                for(int j=0; j<int(_codeLines[i]._vasm.size()); j++)
                 {
                     uint16_t vasmAddress = _codeLines[i]._vasm[j]._address;
                     std::string vasmCode = _codeLines[i]._vasm[j]._code;
@@ -2782,17 +2782,17 @@ namespace Compiler
 
     void discardUnusedLabels(void)
     {
-        for(size_t k=0; k<_equateLabels.size(); k++)
+        for(int k=0; k<int(_equateLabels.size()); k++)
         {
             bool foundLabel = false;
 
-            for(size_t i=0; i<_codeLines.size(); i++)
+            for(int i=0; i<int(_codeLines.size()); i++)
             {
                 // Valid BASIC code
                 if(_codeLines[i]._code.size() >= 2  &&  _codeLines[i]._vasm.size())
                 {
                     // Vasm code
-                    for(size_t j=0; j<_codeLines[i]._vasm.size(); j++)
+                    for(int j=0; j<int(_codeLines[i]._vasm.size()); j++)
                     {
                         if(_codeLines[i]._vasm[j]._code.find(_equateLabels[k]) != std::string::npos)
                         {
@@ -2811,7 +2811,7 @@ namespace Compiler
                                       _equateLabels[k].find("_endif_") != std::string::npos  ||  _equateLabels[k].find("_while_") != std::string::npos ||  _equateLabels[k].find("_wend_") != std::string::npos    ||
                                       _equateLabels[k].find("_repeat_") != std::string::npos ||  _equateLabels[k].find("_next_") != std::string::npos  ||  _equateLabels[k].find("_page_") != std::string::npos);
 
-                for(size_t l=0; l<_output.size(); l++)
+                for(int l=0; l<int(_output.size()); l++)
                 {
                     // Find unused internal label in output and discard it
                     if(foundInternal  &&  _output[l].find(_equateLabels[k]) != std::string::npos)
